LayoutTests/fast/selectors/attribute-exact-value-match-is-ascii-case-insensitive-expected.txt

When matching attributes case insensitively, it should be ASCII case insensitive. This test verifies the behavior for exact value matching (e.g. [a="b"])

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing "[data-attribute=WebKit-É]"
PASS document.querySelectorAll('#test-root [data-attribute=WebKit-É]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute=WebKit-É]')[0].id is "target1"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute=webkit-É]"
PASS document.querySelectorAll('#test-root [data-attribute=webkit-É]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute=webkit-É]')[0].id is "target2"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute=WEBKIT-É]"
PASS document.querySelectorAll('#test-root [data-attribute=WEBKIT-É]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute=WEBKIT-É]')[0].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute=WebKit-é]"
PASS document.querySelectorAll('#test-root [data-attribute=WebKit-é]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute=WebKit-é]')[0].id is "target4"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute=webkit-é]"
PASS document.querySelectorAll('#test-root [data-attribute=webkit-é]').length is 1
PASS document.querySelectorAll('#test-root [data-attribute=webkit-é]')[0].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[data-attribute=WebKit-É i]"
PASS document.querySelectorAll('#test-root [data-attribute=WebKit-É i]').length is 3
PASS document.querySelectorAll('#test-root [data-attribute=WebKit-É i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [data-attribute=WebKit-É i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [data-attribute=WebKit-É i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute=webkit-É i]"
PASS document.querySelectorAll('#test-root [data-attribute=webkit-É i]').length is 3
PASS document.querySelectorAll('#test-root [data-attribute=webkit-É i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [data-attribute=webkit-É i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [data-attribute=webkit-É i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute=WEBKIT-É i]"
PASS document.querySelectorAll('#test-root [data-attribute=WEBKIT-É i]').length is 3
PASS document.querySelectorAll('#test-root [data-attribute=WEBKIT-É i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [data-attribute=WEBKIT-É i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [data-attribute=WEBKIT-É i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[data-attribute=WebKit-é i]"
PASS document.querySelectorAll('#test-root [data-attribute=WebKit-é i]').length is 2
PASS document.querySelectorAll('#test-root [data-attribute=WebKit-é i]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [data-attribute=WebKit-é i]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[data-attribute=webkit-é i]"
PASS document.querySelectorAll('#test-root [data-attribute=webkit-é i]').length is 2
PASS document.querySelectorAll('#test-root [data-attribute=webkit-é i]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [data-attribute=webkit-é i]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[multiple=WEBKIT-é]"
PASS document.querySelectorAll('#test-root [multiple=WEBKIT-é]').length is 3
PASS document.querySelectorAll('#test-root [multiple=WEBKIT-é]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple=WEBKIT-é]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple=WEBKIT-é]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple=WebKit-é]"
PASS document.querySelectorAll('#test-root [multiple=WebKit-é]').length is 3
PASS document.querySelectorAll('#test-root [multiple=WebKit-é]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple=WebKit-é]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple=WebKit-é]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple=Webkit-é]"
PASS document.querySelectorAll('#test-root [multiple=Webkit-é]').length is 3
PASS document.querySelectorAll('#test-root [multiple=Webkit-é]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple=Webkit-é]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple=Webkit-é]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple=webkit-É]"
PASS document.querySelectorAll('#test-root [multiple=webkit-É]').length is 2
PASS document.querySelectorAll('#test-root [multiple=webkit-É]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [multiple=webkit-É]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[multiple=webKit-É]"
PASS document.querySelectorAll('#test-root [multiple=webKit-É]').length is 2
PASS document.querySelectorAll('#test-root [multiple=webKit-É]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [multiple=webKit-É]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[multiple=WEBKIT-é i]"
PASS document.querySelectorAll('#test-root [multiple=WEBKIT-é i]').length is 3
PASS document.querySelectorAll('#test-root [multiple=WEBKIT-é i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple=WEBKIT-é i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple=WEBKIT-é i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple=WebKit-é i]"
PASS document.querySelectorAll('#test-root [multiple=WebKit-é i]').length is 3
PASS document.querySelectorAll('#test-root [multiple=WebKit-é i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple=WebKit-é i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple=WebKit-é i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple=Webkit-é i]"
PASS document.querySelectorAll('#test-root [multiple=Webkit-é i]').length is 3
PASS document.querySelectorAll('#test-root [multiple=Webkit-é i]')[0].id is "target1"
PASS document.querySelectorAll('#test-root [multiple=Webkit-é i]')[1].id is "target2"
PASS document.querySelectorAll('#test-root [multiple=Webkit-é i]')[2].id is "target3"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(255, 0, 0)"

Testing "[multiple=webkit-É i]"
PASS document.querySelectorAll('#test-root [multiple=webkit-É i]').length is 2
PASS document.querySelectorAll('#test-root [multiple=webkit-É i]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [multiple=webkit-É i]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

Testing "[multiple=webKit-É i]"
PASS document.querySelectorAll('#test-root [multiple=webKit-É i]').length is 2
PASS document.querySelectorAll('#test-root [multiple=webKit-É i]')[0].id is "target4"
PASS document.querySelectorAll('#test-root [multiple=webKit-É i]')[1].id is "target5"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[0]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[1]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[2]).backgroundColor is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[3]).backgroundColor is "rgb(10, 100, 200)"
PASS getComputedStyle(document.querySelectorAll("#test-root *")[4]).backgroundColor is "rgb(10, 100, 200)"

PASS successfullyParsed is true

TEST COMPLETE
